Build an aushelper v2.5 to ship to users of Firefox 52.
Firefox 52 shipped with this system add-on built-in but due to a bug
in the update process, users who might had already received an earlier
version in the past were not getting upgraded to this current version.

This commit restructures the aushelper folder as v1.5 that existed before
and this new v2.5.